+++ test/libsolidity/syntaxTests/inlineAssembly/extcodehash_as_variable_post_constantinople.sol
@@ -1,0 +1,9 @@
+contract c {
+	function f() public view {
+		uint extcodehash;
+		extcodehash;
+		assembly { pop(extcodehash(0)) }
+	}
+}
+// ====
+// EVMVersion: >=constantinople