+++ test/libsolidity/syntaxTests/inlineAssembly/shadowing/variable_by_opcode.sol
@@ -9,6 +9,5 @@ contract C {
     }
 }
 // ----
-// Warning 8261: (109-119): Variable is shadowed in inline assembly by an instruction of the same name
 // Warning 2072: (52-62): Unused local variable.
 // Warning 2072: (109-119): Unused local variable.